requirements.txt
jira==3.5.1
python-dotenv==1.0.0
anthropic>=0.9.0
faker==22.6.0
python-dateutil==2.8.2 